--- app/(app)/onboarding/index.css
@@ -1,28 +1,0 @@
-/* index.css */
-@import 'tailwindcss/base';
-@import 'tailwindcss/components';
-@import 'tailwindcss/utilities';
-
-@import 'daisyui/dist/full.css';
-
-.custom-font {
-  font-family: Arial, sans-serif;
-}
-
-.container {
-  text-align: center;
-  margin-top: 80px;
-  margin-left: 30px;
-}
-
-.custom-button {
-  @apply bg-purple-600 text-white rounded-lg border-none py-2 px-6 text-lg cursor-pointer mb-5;
-}
-
-.custom-button:hover {
-  @apply bg-purple-700;
-}
-
-h1 {
-  font-size: 96px;
-  color: #231
